ToDo app
This is a simple application where the user can submit a todo list by typing them in and as they clear the tasks they can start scratching. Initially each ToDo appears green colored and when cleared changes to red color with a line through. 
If the user submits a ToDo then decides to give up on it, they can delete it by right clicking